auth0 navigation fixed
